Renamed the connected container components from *-container.js to index.js.
